refactor(core): save stockholders using AutoSaveStartTask
