+++ core/src/main/java/net/okocraft/box/core/listener/StockHolderListener.java
@@ -1,7 +1,8 @@
-package net.okocraft.box.core.task;
+package net.okocraft.box.core.listener;
 
 import com.github.siroshun09.event4j.handlerlist.Key;
 import net.okocraft.box.api.BoxProvider;
+import net.okocraft.box.api.event.general.AutoSaveStartEvent;
 import net.okocraft.box.api.event.stock.StockDecreaseEvent;
 import net.okocraft.box.api.event.stock.StockEvent;
 import net.okocraft.box.api.event.stock.StockIncreaseEvent;
@@ -9,54 +10,35 @@ import net.okocraft.box.api.event.stock.StockSaveEvent;
 import net.okocraft.box.api.event.stock.StockSetEvent;
 import net.okocraft.box.api.model.stock.UserStockHolder;
 import net.okocraft.box.api.model.user.BoxUser;
-import net.okocraft.box.core.config.Settings;
 import net.okocraft.box.core.message.ErrorMessages;
-import net.okocraft.box.core.storage.Storage;
-import net.okocraft.box.core.util.executor.InternalExecutors;
 import org.bukkit.Bukkit;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.ScheduledFuture;
-import java.util.concurrent.TimeUnit;
 import java.util.logging.Level;
 
-public class ModifiedStockHolderSaveTask {
+public class StockHolderListener {
 
     private final Key listenerKey = Key.of(getClass().getSimpleName());
     private final List<UserStockHolder> modifiedStockHolders = new ArrayList<>();
 
-    private final Storage storage;
-
-    private ScheduledFuture<?> task;
-
-    public ModifiedStockHolderSaveTask(@NotNull Storage storage) {
-        this.storage = storage;
-    }
-
-    public void start() {
+    public void register() {
         var eventBus = BoxProvider.get().getEventBus();
 
-        eventBus.getHandlerList(StockSetEvent.class).subscribe(listenerKey, this::processEvent);
-        eventBus.getHandlerList(StockIncreaseEvent.class).subscribe(listenerKey, this::processEvent);
-        eventBus.getHandlerList(StockDecreaseEvent.class).subscribe(listenerKey, this::processEvent);
+        eventBus.getHandlerList(StockSetEvent.class).subscribe(listenerKey, this::queueStockHolder);
+        eventBus.getHandlerList(StockIncreaseEvent.class).subscribe(listenerKey, this::queueStockHolder);
+        eventBus.getHandlerList(StockDecreaseEvent.class).subscribe(listenerKey, this::queueStockHolder);
 
-        var scheduler = InternalExecutors.newSingleThreadScheduler("Auto-Save-Scheduler");
-        long interval = BoxProvider.get().getConfiguration().get(Settings.STOCK_DATA_SAVE_INTERVAL);
-        task = scheduler.scheduleAtFixedRate(this::runTask, interval, interval, TimeUnit.SECONDS);
+        eventBus.getHandlerList(AutoSaveStartEvent.class).subscribe(listenerKey, this::saveModifiedStockHolders);
     }
 
-    public void stop() {
+    public void unregister() {
         BoxProvider.get().getEventBus().unsubscribeAll(listenerKey);
-
-        if (task != null) {
-            task.cancel(true);
-        }
     }
 
-    private void processEvent(@NotNull StockEvent event) {
+    private void queueStockHolder(@NotNull StockEvent event) {
         if (event.isUserStockHolder()) {
             var userStockHolder = event.getUserStockHolder();
 
@@ -66,7 +48,7 @@ public class ModifiedStockHolderSaveTask {
         }
     }
 
-    private void runTask() {
+    private void saveModifiedStockHolders(@NotNull AutoSaveStartEvent task) {
         if (modifiedStockHolders.isEmpty()) {
             return;
         }
@@ -75,15 +57,12 @@ public class ModifiedStockHolderSaveTask {
 
         modifiedStockHolders.clear();
 
-        copied.stream()
-                .filter(UserStockHolder::isOnline)
-                .forEach(this::saveUserStockHolder);
-
+        copied.stream().filter(UserStockHolder::isOnline).forEach(this::save);
     }
 
-    private void saveUserStockHolder(@NotNull UserStockHolder userStockHolder) {
+    private void save(@NotNull UserStockHolder userStockHolder) {
         try {
-            storage.getStockStorage().saveUserStockHolder(userStockHolder);
+            BoxProvider.get().getStockManager().saveUserStock(userStockHolder).join();
             BoxProvider.get().getEventBus().callEvent(new StockSaveEvent(userStockHolder));
         } catch (Exception e) {
             Optional.of(userStockHolder.getUser())
